+++ cider-mode.el
@@ -1032,9 +1032,7 @@ property."
       (progn
         (setq-local sesman-system 'CIDER)
         (cider-eldoc-setup)
-        (make-local-variable 'completion-at-point-functions)
-        (add-to-list 'completion-at-point-functions
-                     #'cider-complete-at-point)
+        (add-hook 'completion-at-point-functions #'cider-complete-at-point nil t)
         (font-lock-add-keywords nil cider--static-font-lock-keywords)
         (cider-refresh-dynamic-font-lock)
         (font-lock-add-keywords nil cider--reader-conditionals-font-lock-keywords)
@@ -1050,6 +1048,7 @@ property."
           (setq-local clojure-get-indent-function #'cider--get-symbol-indent))
         (setq-local clojure-expected-ns-function #'cider-expected-ns)
         (setq next-error-function #'cider-jump-to-compilation-error))
+    ;; Mode cleanup
     (mapc #'kill-local-variable '(completion-at-point-functions
                                   next-error-function
                                   x-gtk-use-system-tooltips
@@ -1059,7 +1058,8 @@ property."
     (font-lock-add-keywords nil cider--reader-conditionals-font-lock-keywords)
     (font-lock-remove-keywords nil cider--dynamic-font-lock-keywords)
     (font-lock-remove-keywords nil cider--static-font-lock-keywords)
-    (cider--font-lock-flush)))
+    (cider--font-lock-flush)
+    (remove-hook 'completion-at-point-functions #'cider-complete-at-point t)))
 
 (defun cider-set-buffer-ns (ns)
   "Set this buffer's namespace to NS and refresh font-locking."